environment file updated with postgres ip and port
