+++ dummy_data/clear_data.py
@@ -3,15 +3,29 @@ import math
 
 import psycopg2
 from psycopg2 import sql
+import os
+from dotenv import load_dotenv
 
+# Specify the path to the .env file
+dotenv_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env')
+
+# Load the environment variables from the .env file
+load_dotenv(dotenv_path)
+
+# Access environment variables
+pg_host = os.getenv('POSTGRES_HOST')
+pg_port = os.getenv('POSTGRES_PORT')
+
+# print(f"postgres host: {pg_host}")
+# print(f"postgres port: {pg_port}")
 
 
 
 
 
 conn = psycopg2.connect(
-    host="0.0.0.0", 
-    port="45432",
+    host=pg_host, 
+    port=pg_port,
     database="istsos",
     user="admin",
     password="admin"
@@ -20,9 +34,6 @@ conn = psycopg2.connect(
 def clear():
 
         
-
-
-
 
     # SQL statement for updating location geometry and updating id sequence number of 8 columns
     clear_table = f"""
