+++ dummy_data/data/Sensor.csv
@@ -1,2 +1,2 @@
 id,name,description,encodingType,metadata,properties
-2,Light Sensor_2,Sensor is a Light Sensor_2,application/pdf,{},{}
+2,Temperature Sensor_2,Sensor is a Temperature Sensor_2,application/pdf,{},{}
