+++ dummy_data/data/Thing.csv
@@ -1,2 +1,2 @@
 id,name,description,properties,location_id
-2,Motion Sensor_2,A sensor that measures the Motion Sensor_2 in a room,"{""model"": ""Model-2"", ""manufacturer"": ""Manufacturer-2""}",46
+2,Light Sensor_2,A sensor that measures the Light Sensor_2 in a room,"{""model"": ""Model-2"", ""manufacturer"": ""Manufacturer-2""}",92
