+++ dummy_data/requirements.txt
@@ -1,2 +1,3 @@
 psycopg2
 PyYAML
+python-dotenv
